Brian

+++ phase3/libuser.h
@@ -6,10 +6,10 @@
 #define _LIBUSER_H
 
 // Phase 3 -- User Function Prototypes
-extern int  Spawn(char *name, int (*func)(char *), char *arg, int stack_size,
-                  int priority, int *pid);
+extern int  Spawn(char *name, int (*func)(char *), char *arg, long stack_size,
+                  long priority, int *pid);
 extern int  Wait(int *pid, int *status);
-extern void Terminate(int status);
+extern void Terminate(long status);
 extern void GetTimeofDay(int *tod);
 extern void CPUTime(int *cpu);
 extern void GetPID(int *pid);
